+++ src/app/lastfm/manager.rs
@@ -3,9 +3,7 @@ use rustfm_scrobble::{Scrobble, Scrobbler};
 use serde::Deserialize;
 use std::path::Path;
 
-use crate::app::index::Index;
-use crate::db::DB;
-use crate::user;
+use crate::app::{index::Index, user};
 
 const LASTFM_API_KEY: &str = "02b96c939a2b451c31dfd67add1f696e";
 const LASTFM_API_SECRET: &str = "0f25a80ceef4b470b5cb97d99d4b3420";
@@ -42,30 +40,34 @@ struct AuthResponse {
 }
 
 pub struct Manager {
-	db: DB,
 	index: Index,
+	user_manager: user::Manager,
 }
 
 impl Manager {
-	pub fn new(db: DB, index: Index) -> Self {
-		Self { db, index }
+	pub fn new(index: Index, user_manager: user::Manager) -> Self {
+		Self {
+			index,
+			user_manager,
+		}
 	}
 
 	pub fn link(&self, username: &str, token: &str) -> Result<()> {
 		let mut scrobbler = Scrobbler::new(LASTFM_API_KEY.into(), LASTFM_API_SECRET.into());
 		let auth_response = scrobbler.authenticate_with_token(token)?;
 
-		user::lastfm_link(&self.db, username, &auth_response.name, &auth_response.key)
+		self.user_manager
+			.lastfm_link(username, &auth_response.name, &auth_response.key)
 	}
 
 	pub fn unlink(&self, username: &str) -> Result<()> {
-		user::lastfm_unlink(&self.db, username)
+		self.user_manager.lastfm_unlink(username)
 	}
 
 	pub fn scrobble(&self, username: &str, track: &Path) -> Result<()> {
 		let mut scrobbler = Scrobbler::new(LASTFM_API_KEY.into(), LASTFM_API_SECRET.into());
 		let scrobble = self.scrobble_from_path(track)?;
-		let auth_token = user::get_lastfm_session_key(&self.db, username)?;
+		let auth_token = self.user_manager.get_lastfm_session_key(username)?;
 		scrobbler.authenticate_with_session_key(&auth_token);
 		scrobbler.scrobble(&scrobble)?;
 		Ok(())
@@ -74,7 +76,7 @@ impl Manager {
 	pub fn now_playing(&self, username: &str, track: &Path) -> Result<()> {
 		let mut scrobbler = Scrobbler::new(LASTFM_API_KEY.into(), LASTFM_API_SECRET.into());
 		let scrobble = self.scrobble_from_path(track)?;
-		let auth_token = user::get_lastfm_session_key(&self.db, username)?;
+		let auth_token = self.user_manager.get_lastfm_session_key(username)?;
 		scrobbler.authenticate_with_session_key(&auth_token);
 		scrobbler.now_playing(&scrobble)?;
 		Ok(())
