+++ src/app/mod.rs
@@ -3,4 +3,5 @@ pub mod index;
 pub mod lastfm;
 pub mod playlists;
 pub mod thumbnails;
+pub mod user;
 pub mod vfs;
