+++ src/app/user/mod.rs
@@ -1,0 +1,3 @@
+mod manager;
+
+pub use manager::*;
